Refactor hazard selection and visibility hooks
